Eliminate the need to patch ssh/client.go
Instead of adding 2 to the wait group for every host we will connect to,
we now add only 1 in the beginning and another as soon as we have a host
key to print. That way, every host connection will end in an error
(even if host keys were retrieved, we cannot authenticate) and we simply
do not care.

Note that if we wanted to print error messages, we still would need to
throw a special error in our host key callback in order to distinguish
"real" errors from "we threw an error to abort the connection attempt"